requirement.txt

numpy
pandas
tqdm
lmdb
opencv-python
pillow
tensorboardx
wandb

